requirements.txt
opencv-python
numpy
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
mediapipe
python-dotenv
pyserial